+++ lib/src/controls/inputs/buttons/hyperlink_button.dart
@@ -1,8 +1,5 @@
 import 'package:fluent_ui/fluent_ui.dart';
 
-@Deprecated('Deprecated in 4.4.3. Use HypelinkButton instead')
-typedef TextButton = HyperlinkButton;
-
 /// A borderless button with mainly text-based content
 ///
 /// {@macro fluent_ui.buttons.base}